transactions/requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.1
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
pytest>=7.4.0